rename the v2 directory to circuitv2
Otherwise, we'll run into problems if we ever want to release a (semantic)
version v2 of this package.
