+++ circuitv2/client/reservation.go
@@ -5,9 +5,9 @@ import (
 	"fmt"
 	"time"
 
-	pbv2 "github.com/libp2p/go-libp2p-circuit/v2/pb"
-	"github.com/libp2p/go-libp2p-circuit/v2/proto"
-	"github.com/libp2p/go-libp2p-circuit/v2/util"
+	pbv2 "github.com/libp2p/go-libp2p-circuit/circuitv2/pb"
+	"github.com/libp2p/go-libp2p-circuit/circuitv2/proto"
+	"github.com/libp2p/go-libp2p-circuit/circuitv2/util"
 
 	"github.com/libp2p/go-libp2p-core/host"
 	"github.com/libp2p/go-libp2p-core/peer"
