+++ circuitv2/test/e2e_test.go
@@ -9,8 +9,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/libp2p/go-libp2p-circuit/v2/client"
-	"github.com/libp2p/go-libp2p-circuit/v2/relay"
+	"github.com/libp2p/go-libp2p-circuit/circuitv2/client"
+	"github.com/libp2p/go-libp2p-circuit/circuitv2/relay"
 
 	"github.com/libp2p/go-libp2p-core/crypto"
 	"github.com/libp2p/go-libp2p-core/host"
