+++ circuitv2/util/ma.go
